+++ diesel/src/pg/expression/operators.rs
@@ -1,5 +1,6 @@
 use crate::expression::expression_types::NotSelectable;
 use crate::pg::Pg;
+use crate::sql_types::{Bigint, Inet};
 
 infix_operator!(IsDistinctFrom, " IS DISTINCT FROM ", backend: Pg);
 infix_operator!(IsNotDistinctFrom, " IS NOT DISTINCT FROM ", backend: Pg);
@@ -12,3 +13,10 @@ infix_operator!(SimilarTo, " SIMILAR TO ", backend: Pg);
 infix_operator!(NotSimilarTo, " NOT SIMILAR TO ", backend: Pg);
 postfix_operator!(NullsFirst, " NULLS FIRST", NotSelectable, backend: Pg);
 postfix_operator!(NullsLast, " NULLS LAST", NotSelectable, backend: Pg);
+infix_operator!(ContainsNet, " >> ", backend: Pg);
+infix_operator!(ContainsNetLoose, " >>= ", backend: Pg);
+infix_operator!(IsContainedByNet, " << ", backend: Pg);
+infix_operator!(IsContainedByNetLoose, " <<= ", backend: Pg);
+infix_operator!(AndNet, " & ", Inet, backend: Pg);
+infix_operator!(OrNet, " | ", Inet, backend: Pg);
+infix_operator!(DifferenceNet, " - ", Bigint, backend: Pg);