+++ diesel/src/query_dsl/locking_dsl.rs
@@ -1,5 +1,9 @@
+use crate::expression::TypedExpressionType;
+use crate::expression::ValidGrouping;
 use crate::query_builder::AsQuery;
+use crate::query_builder::SelectStatement;
 use crate::query_source::Table;
+use crate::Expression;
 
 /// Methods related to locking select statements
 ///
@@ -21,10 +25,11 @@ pub trait LockingDsl<Lock> {
 
 impl<T, Lock> LockingDsl<Lock> for T
 where
-    T: Table + AsQuery,
-    T::Query: LockingDsl<Lock>,
+    T: Table + AsQuery<Query = SelectStatement<T>>,
+    T::DefaultSelection: Expression<SqlType = T::SqlType> + ValidGrouping<()>,
+    T::SqlType: TypedExpressionType,
 {
-    type Output = <T::Query as LockingDsl<Lock>>::Output;
+    type Output = <SelectStatement<T> as LockingDsl<Lock>>::Output;
 
     fn with_lock(self, lock: Lock) -> Self::Output {
         self.as_query().with_lock(lock)