+++ CHANGELOG.md
@@ -50,6 +50,10 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 * Added `#[diesel(serialize_as)]` analogous to `#[diesel(deserialize_as)]`. This allows
   customization of the serialization behaviour of `Insertable` structs.
 
+* Added support for `GROUP BY` clauses
+
+* Added support for `UNION`, `UNION ALL`, `INTERSECT`, `INTERSECT ALL`, `EXCEPT`, `EXCEPT ALL` clauses
+
 ### Removed
 
 * All previously deprecated items have been removed.
@@ -58,6 +62,7 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 * Support for `pq-sys` < 0.4.0 has been removed.
 * Support for `mysqlclient-sys` < 0.2.0 has been removed.
 * Support for `time` types has been removed.
+* Support for `chrono` < 0.4.19 has been removed.
 * The `NonNull` for sql types has been removed in favour of the new `SqlType` trait.
 
 * `no_arg_sql_function!` has been deprecated without replacement.